gemini_query/requirements.txt
google-genai>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
